:sparkles: ToDo 인가 적용 - 로그인 한 사용자만 생성 가능 - 작성자만 수정, 삭제 가능

+++ src/main/kotlin/com/teamsparta/todo/domain/todo/service/ToDoServiceImpl.kt
@@ -1,8 +1,12 @@
 package com.teamsparta.todo.domain.todo.service
 
 import com.teamsparta.todo.domain.todo.dto.*
+import com.teamsparta.todo.domain.todo.model.ToDo
 import com.teamsparta.todo.domain.todo.repository.ToDoRepository
+import com.teamsparta.todo.domain.user.repository.UserRepository
 import com.teamsparta.todo.exception.ModelNotFoundException
+import com.teamsparta.todo.exception.UnauthorizedException
+import com.teamsparta.todo.infra.auth.AuthUser
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort
@@ -12,14 +16,16 @@ import org.springframework.transaction.annotation.Transactional
 
 @Service
 class ToDoServiceImpl(
-    private val toDoRepository: ToDoRepository
+    private val toDoRepository: ToDoRepository,
+    private val userRepository: UserRepository
 ) : ToDoService {
 
     private val pageSize = 5
 
     @Transactional
-    override fun createToDo(request: CreateToDoRequest): ToDoResponse {
-        return toDoRepository.save(request.toEntity()).toResponse()
+    override fun createToDo(loginUser: AuthUser, request: CreateToDoRequest): ToDoResponse {
+        val user = userRepository.findByIdOrNull(loginUser.id) ?: throw ModelNotFoundException("user", loginUser.id)
+        return toDoRepository.save(request.toEntity(user)).toResponse()
     }
 
     override fun getAllToDos(sort: String, cursor: Long): SliceResponse<ToDoResponse> {
@@ -66,21 +72,24 @@ class ToDoServiceImpl(
     }
 
     @Transactional
-    override fun deleteToDo(todoId: Long) {
+    override fun deleteToDo(todoId: Long, loginUser: AuthUser) {
         val todo = toDoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("ToDo", todoId)
+        checkAuthorization(todo, loginUser)
         toDoRepository.delete(todo)
     }
 
     @Transactional
-    override fun updateToDo(toDoId: Long, request: UpdateToDoRequest): ToDoResponse {
+    override fun updateToDo(toDoId: Long, loginUser: AuthUser, request: UpdateToDoRequest): ToDoResponse {
         val todo = toDoRepository.findByIdOrNull(toDoId) ?: throw ModelNotFoundException("ToDo", toDoId)
+        checkAuthorization(todo, loginUser)
         todo.update(request)
         return todo.toResponse()
     }
 
     @Transactional
-    override fun updateToDoStatus(toDoId: Long, status: Boolean): ToDoResponse {
+    override fun updateToDoStatus(toDoId: Long, loginUser: AuthUser, status: Boolean): ToDoResponse {
         val todo = toDoRepository.findByIdOrNull(toDoId) ?: throw ModelNotFoundException("ToDo", toDoId)
+        checkAuthorization(todo, loginUser)
         todo.changeStatus(status)
         return todo.toResponse()
     }
@@ -88,5 +97,10 @@ class ToDoServiceImpl(
     private fun getDirection(sort: String) = when (sort) {
         "asc" -> Sort.Direction.ASC
         else -> Sort.Direction.DESC
+    }
+
+    private fun checkAuthorization(todo: ToDo, loginUser: AuthUser) {
+        val user = userRepository.findByIdOrNull(loginUser.id) ?: throw ModelNotFoundException("User", loginUser.id)
+        if (user.id != todo.createdBy.id) throw UnauthorizedException("권한이 없습니다.")
     }
 }
